requirements.txt
accelerate==0.31.0
autocorrect==2.6.1
bitsandbytes==0.43.1
datasets==2.19.2
evaluate==0.4.2
gensim==4.3.2
gputil==1.4.0
huggingface-hub==0.23.3
icdiff==2.0.7
keras==3.3.3
nltk==3.8.1
numpy==1.26.4
peft==0.11.1
psutil==5.9.8
pyarrow==16.1.0
python-dotenv==1.0.1
pyspellchecker==0.8.1
pyyaml==6.0.1
scipy==1.12
sentence-transformers==3.0.1
tensorflow==2.16.1
tensorflow-hub==0.16.1
textattack==0.3.10
tokenizers==0.19.1
torch==2.3.1
transformers==4.41.2
trl==0.9.6
typed-argument-parser==1.10.0
